+++ aries-site/src/examples/components/avatar/index.js
@@ -1,8 +1,7 @@
 export * from './AvatarAccessibilityExample';
 export * from './AvatarCardExampleDo';
 export * from './AvatarCardExampleDont';
-export * from './AvatarDataTableExampleDo';
-export * from './AvatarDataTableExampleDont';
+export * from './AvatarDataTableExample';
 export * from './AvatarExample';
 export * from './AvatarPageHeaderExample';
 export * from './AvatarSizePreviewExample';
